@@ -3,11 +3,12 @@
   "version": "0.0.1",
   "private": true,
   "dependencies": {
-    "@capacitor/android": "^2.1.0",
-    "@capacitor/core": "2.1.0",
-    "@capacitor/ios": "^2.1.0",
-    "@ionic/react": "^5.0.7",
-    "@ionic/react-router": "^5.0.7",
+    "@capacitor/android": "7.3.0",
+    "@capacitor/browser": "^7.0.1",
+    "@capacitor/core": "7.3.0",
+    "@capacitor/ios": "7.3.0",
+    "@ionic/react": "^8.6.0",
+    "@ionic/react-router": "^8.6.0",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.4.0",
     "@testing-library/user-event": "^8.0.3",
@@ -18,13 +19,12 @@
     "@types/react-router": "^5.1.4",
     "@types/react-router-dom": "^5.1.3",
     "date-fns": "^2.14.0",
-    "firebase": "^9.23.0",
+    "firebase": "9.6.11",
     "ionicons": "^5.0.0",
-    "react": "^16.13.0",
-    "react-dom": "^16.13.0",
-    "react-router": "^5.1.2",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "5.0.1",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-router": "^5.3.4",
+    "react-router-dom": "^5.3.4",
     "typescript": "3.8.3"
   },
   "scripts": {
@@ -50,7 +50,10 @@
     ]
   },
   "devDependencies": {
-    "@capacitor/cli": "5.1.0"
+    "@babel/plugin-transform-private-property-in-object": "^7.27.1",
+    "@capacitor/cli": "7.3.0",
+    "vite": "^5.3.1",
+    "@vitejs/plugin-react": "^4.3.1"
   },
   "description": "PinIdeas | Chinmay Bhushan"
 }
